chore: Remove superseded `SortMenu` component (#63186)
* Remove superseded `SortMenu` component

* Rename `SortMenuV2` to `SortMenu`

* Add backwards compatibility for teleport.e
