+++ web/packages/shared/components/UnifiedResources/FilterPanel.tsx
@@ -27,7 +27,7 @@ import Menu from 'design/Menu';
 import { HoverTooltip } from 'design/Tooltip';
 import { ViewMode } from 'gen-proto-ts/teleport/userpreferences/v1/unified_resource_preferences_pb';
 import { MultiselectMenu } from 'shared/components/Controls/MultiselectMenu';
-import { SortItem, SortMenu } from 'shared/components/Controls/SortMenuV2';
+import { SortItem, SortMenu } from 'shared/components/Controls/SortMenu';
 import { ViewModeSwitch } from 'shared/components/Controls/ViewModeSwitch';
 
 import {
