+++ web/packages/teleport/src/SessionRecordings/list/state.ts
@@ -27,7 +27,7 @@ import {
 } from 'react';
 import { useHistory } from 'react-router';
 
-import { SortOrder } from 'shared/components/Controls/SortMenuV2';
+import { SortOrder } from 'shared/components/Controls/SortMenu';
 
 import {
   EventRange,
